Fixed logo on reset password pages fixed the run importer button on the user management settings
Added root and qulaity profile selection for movies #1517
Added the Sonarr Cacher #1513

Refactored what we do to tv searches to use the rules engine
Cache a few more things to speed some searches up

+++ src/Ombi.TheMovieDbApi/Models/MovieResponse.cs
@@ -45,8 +45,8 @@ namespace Ombi.TheMovieDbApi.Models
         public ProductionCompanies[] production_companies { get; set; }
         public ProductionCountries[] production_countries { get; set; }
         public string release_date { get; set; }
-        public int revenue { get; set; }
-        public int runtime { get; set; }
+        public float revenue { get; set; }
+        public float runtime { get; set; }
         public SpokenLanguages[] spoken_languages { get; set; }
         public string status { get; set; }
         public string tagline { get; set; }
